+++ packages/address-picker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - @lendi-ui/address-picker
 
+## [2.0.3] 2019-10-21 [HUB-639](https://creditandfinance.atlassian.net/browse/HUB-639)
+### Changed
+- Replaced generic ERROR with ZERO_RESULTS error from google.
+
 ## [2.0.2] 2019-10-08 [HUB-586](https://creditandfinance.atlassian.net/browse/HUB-586)
 ### Updated
 - Bumped the version of button to ^8 and changed the button type inside the modal to button (it was previously submit by default).